# shared/api-contracts/openapi.yaml
...
openapi: "3.1.0"
info:
  title: "TTBS Customer API"
  description: "Customer-facing API contract for the train ticket booking system."
  version: "1.0"
servers:
- url: "/api"
  description: "Versioned customer API root"
security:
- bearerAuth: []
...
paths:
...
  /api/{version}/auth/register:
    post:
      tags:
      - "Authentication"
      summary: "Register a customer account"
      operationId: "register"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
        required: true
      responses:
        "201":
          description: "Customer account created"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/JsendResponseObject"
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          description: "Invalid registration payload"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/JsendFailResponse"
        "409":
          description: "Email address already exists"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/JsendFailResponse"
        "500":
          description: "Unexpected technical failure."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JsendErrorResponse"
  /api/{version}/auth/refresh:
    post:
      tags:
      - "Authentication"
      summary: "Rotate access and refresh tokens"
      operationId: "refreshToken"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenRequest"
        required: true
      responses:
        "200":
          description: "Token pair rotated"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/JsendResponseObject"
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResultResponse"
        "400":
          description: "Invalid refresh-token payload"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/JsendFailResponse"
        "401":
          description: "Refresh token is invalid or expired"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/JsendFailResponse"
        "500":
          description: "Unexpected technical failure."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JsendErrorResponse"
  /api/{version}/auth/logout:
    post:
      tags:
      - "Authentication"
      summary: "Revoke the current refresh token"
      operationId: "logout"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenRequest"
        required: true
      responses:
        "200":
          description: "Refresh token revoked"
        "400":
          description: "Invalid refresh-token payload"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/JsendFailResponse"
        "401":
          description: "Refresh token is invalid or expired"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/JsendFailResponse"
        "500":
          description: "Unexpected technical failure."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JsendErrorResponse"
  /api/{version}/auth/login:
    post:
      tags:
      - "Authentication"
      summary: "Authenticate a customer"
      operationId: "login"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
        required: true
      responses:
        "200":
          description: "Authentication successful"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/JsendResponseObject"
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResultResponse"
        "400":
          description: "Invalid login payload"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/JsendFailResponse"
        "401":
          description: "Invalid customer credentials"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/JsendFailResponse"
        "500":
          description: "Unexpected technical failure."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JsendErrorResponse"
  /api/{version}/auth/me:
    get:
      tags:
      - "Authentication"
      summary: "Get the current customer profile"
      operationId: "getAuthenticatedUser"
      responses:
        "200":
          description: "Authenticated customer profile"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/JsendResponseObject"
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "401":
          description: "Authentication required"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/JsendFailResponse"
        "404":
          description: "Customer profile not found"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/JsendFailResponse"
        "500":
          description: "Unexpected technical failure."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JsendErrorResponse"
      security:
      - bearerAuth: []
    delete:
      tags:
      - "Authentication"
      summary: "Soft-delete the current customer account"
      operationId: "deleteAuthenticatedUser"
      responses:
        "200":
          description: "Customer account deleted"
        "401":
          description: "Authentication required"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/JsendFailResponse"
        "404":
          description: "Customer profile not found"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/JsendFailResponse"
        "409":
          description: "Customer still has active bookings"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/JsendFailResponse"
        "500":
          description: "Unexpected technical failure."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JsendErrorResponse"
      security:
      - bearerAuth: []
    patch:
      tags:
      - "Authentication"
      summary: "Update the current customer profile"
      operationId: "updateAuthenticatedUser"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAuthenticatedUserRequest"
        required: true
      responses:
        "200":
          description: "Customer profile updated"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/JsendResponseObject"
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          description: "Invalid update payload"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/JsendFailResponse"
        "401":
          description: "Authentication required"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/JsendFailResponse"
        "404":
          description: "Customer profile not found"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/JsendFailResponse"
        "500":
          description: "Unexpected technical failure."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JsendErrorResponse"
      security:
      - bearerAuth: []
...
components:
  schemas:
    FailData:
      description: "Shared payload used inside JSend fail responses."
      properties:
        message:
          type: "string"
          description: "Human-readable summary of the failure."
        code:
          $ref: "#/components/schemas/ErrorCode"
        errors:
          type: "array"
          items:
            $ref: "#/components/schemas/Violation"
    ErrorCode:
      type: "string"
      description: "Machine-readable error code carried by shared fail responses."
      enum:
      - "VALIDATION_ERROR"
      - "CURSOR_INVALID"
      - "ACCESS_DENIED"
      - "USER_EMAIL_ALREADY_EXISTS"
      - "USER_INVALID_CREDENTIALS"
      - "USER_INVALID_REFRESH_TOKEN"
      - "USER_NOT_FOUND"
      - "USER_CANNOT_BULK_DELETE_SELF"
      - "USER_HAS_ACTIVE_BOOKINGS"
      - "BOOKING_NOT_FOUND"
      - "BOOKING_CANNOT_CONFIRM"
      - "BOOKING_ALREADY_CANCELLED"
      - "TRAIN_NUMBER_ALREADY_EXISTS"
      - "TRAIN_NOT_FOUND"
      - "TRAIN_IN_USE"
      - "SEAT_NOT_FOUND"
      - "SEAT_NUMBER_ALREADY_EXISTS"
      - "SEAT_NOT_AVAILABLE"
      - "SEAT_IN_USE"
      - "SCHEDULED_TRIP_NOT_FOUND"
      - "COACH_NOT_FOUND"
      - "COACH_CAR_NUMBER_ALREADY_EXISTS"
      - "COACH_TRAIN_NOT_FOUND"
      - "COACH_IN_USE"
      - "COACH_CAR_NUMBERS_ALREADY_EXIST"
      - "COACH_DUPLICATE_CAR_NUMBERS_IN_REQUEST"
      - "SEAT_NUMBERS_ALREADY_EXIST"
      - "SEAT_DUPLICATE_SEAT_NUMBERS_IN_REQUEST"
      - "ROUTE_TEMPLATE_NOT_FOUND"
      - "STATION_NOT_FOUND"
      - "STATION_CODE_ALREADY_EXISTS"
      - "STATION_IN_USE"
      - "PAYMENT_NOT_FOUND"
      - "PAYMENT_ALREADY_PROCESSED"
      - "PAYMENT_REFUND_FAILED"
    Violation:
      description: "Field-level validation issue inside a fail response."
      properties:
        field:
          type: "string"
          description: "Request field that failed validation."
        message:
          type: "string"
          description: "Human-readable validation message."
        code:
          $ref: "#/components/schemas/ViolationCode"
    ViolationCode:
      type: "string"
      description: "Machine-readable validation category attached to a field violation."
      enum:
      - "REQUIRED"
      - "INVALID_FORMAT"
      - "TOO_SHORT"
      - "TOO_LONG"
      - "INVALID_VALUE"
    JsendFailResponse:
      type: "object"
      description: "Shared JSend fail envelope used for validation and domain errors."
      properties:
        status:
          type: "string"
          enum:
          - "fail"
        data:
          $ref: "#/components/schemas/FailData"
    JsendErrorResponse:
      type: "object"
      description: "Shared JSend error envelope used for unexpected technical failures."
      properties:
        status:
          type: "string"
          enum:
          - "error"
        message:
          type: "string"
          description: "Human-readable technical failure message."
...
    JsendResponseObject:
      type: "object"
      description: "Runtime JSend response envelope used by the API implementation."
      properties:
        status:
          type: "string"
          description: "JSend status value."
          enum:
          - "success"
          - "fail"
          - "error"
        data:
          description: "Payload returned for the response status."
        message:
          type: "string"
          description: "Human-readable technical error message for error responses."
    RegisterRequest:
      type: "object"
      description: "Customer registration payload."
      properties:
        email:
          type: "string"
          format: "email"
          description: "Email address used to sign in."
          example: "customer@example.com"
          minLength: 1
        password:
          type: "string"
          description: "Account password. The value is accepted on write and never\
            \ returned."
          example: "********"
          maxLength: 2147483647
          minLength: 8
          writeOnly: true
        fullName:
          type: "string"
          description: "Customer full name."
          example: "Nguyen Phuong"
          minLength: 1
      required:
      - "email"
      - "fullName"
      - "password"
    RefreshTokenRequest:
      type: "object"
      description: "Refresh-token payload used to rotate or revoke a session."
      properties:
        refreshToken:
          type: "string"
          description: "Refresh token issued during login. The value is accepted on\
            \ write and never returned."
          example: "redacted-refresh-token"
          minLength: 1
          writeOnly: true
      required:
      - "refreshToken"
    LoginRequest:
      type: "object"
      description: "Customer login payload."
      properties:
        email:
          type: "string"
          format: "email"
          description: "Email address used to sign in."
          example: "customer@example.com"
          minLength: 1
        password:
          type: "string"
          description: "Account password. The value is accepted on write and never\
            \ returned."
          example: "********"
          minLength: 1
          writeOnly: true
      required:
      - "email"
      - "password"
    JsonNullableString:
      type: "object"
      properties:
        present:
          type: "boolean"
        undefined:
          type: "boolean"
    UpdateAuthenticatedUserRequest:
      type: "object"
      description: "Patch payload for the authenticated customer profile. Omit a field\
        \ to leave it unchanged."
      properties:
        fullName:
          $ref: "#/components/schemas/JsonNullableString"
          description: "Updated full name."
          example: "Nguyen Phuong"
        email:
          $ref: "#/components/schemas/JsonNullableString"
          description: "Updated email address."
          example: "customer@example.com"
        phone:
          $ref: "#/components/schemas/JsonNullableString"
          description: "Updated phone number."
          example: "+84901234567"
        dateOfBirth:
          type: "string"
          format: "date"
          description: "Updated date of birth."
        gender:
          $ref: "#/components/schemas/JsonNullableString"
          description: "Updated self-declared gender label."
          example: "female"
        idDocumentNumber:
          $ref: "#/components/schemas/JsonNullableString"
          description: "Updated government-issued identity document number."
          example: "redacted-id-document"
          writeOnly: true
        addressLine:
          $ref: "#/components/schemas/JsonNullableString"
          description: "Updated address line."
          example: "redacted-address"
      required:
      - "email"
      - "fullName"
...
    UserResponse:
      description: "Customer profile returned by authenticated user endpoints."
      properties:
        id:
          type: "string"
          format: "uuid"
          description: "Customer identifier."
          readOnly: true
        email:
          type: "string"
          description: "Customer email address."
          example: "customer@example.com"
        fullName:
          type: "string"
          description: "Customer full name."
          example: "Nguyen Phuong"
        phone:
          type: "string"
          description: "Customer phone number."
          example: "+84901234567"
        dateOfBirth:
          type: "string"
          format: "date"
          description: "Customer date of birth."
        gender:
          type: "string"
          description: "Customer self-declared gender label."
          example: "female"
        idDocumentNumber:
          type: "string"
          description: "Government-issued identity document number on file."
          example: "redacted-id-document"
        addressLine:
          type: "string"
          description: "Primary address line on file."
          example: "redacted-address"
        role:
          type: "string"
          description: "Customer role granted by the system."
          example: "CUSTOMER"
        createdAt:
          type: "string"
          format: "date-time"
          description: "Profile creation timestamp."
          readOnly: true
    LoginResultResponse:
      description: "Authentication result containing rotated tokens and the authenticated\
        \ customer profile."
      properties:
        accessToken:
          type: "string"
          description: "Short-lived bearer access token."
          example: "redacted-access-token"
          readOnly: true
        refreshToken:
          type: "string"
          description: "Refresh token used to rotate the session."
          example: "redacted-refresh-token"
          readOnly: true
        user:
          $ref: "#/components/schemas/UserResponse"
...
  securitySchemes:
    bearerAuth:
      type: "http"
      scheme: "bearer"
      bearerFormat: "JWT"